[fix] dashboard廃止、myに統一、コンテンツごとにページ、クエリを分ける
