+++ app/routes/($lang)._main._index/_components/home-user-navigation-menu.tsx
@@ -158,12 +158,12 @@ export const HomeUserNavigationMenu = (props: Props) => {
             label="マイページ"
           />
           <MenuItemLink
-            href={"/dashboard"}
+            href={"/my"}
             icon={<SquareKanbanIcon className="mr-2 inline-block w-4" />}
             label="ダッシュボード"
           />
           <MenuItemLink
-            href={"/dashboard/posts"}
+            href={"/my/posts"}
             icon={<SquareKanbanIcon className="mr-2 inline-block w-4" />}
             label="自分の作品"
           />
