+++ app/routes/($lang)._main.users.$user/_components/user-albums-contents.tsx
@@ -7,7 +7,7 @@ import { ResponsivePagination } from "@/_components/responsive-pagination"
 import { AuthContext } from "@/_contexts/auth-context"
 import type { IntrospectionEnum } from "@/_lib/introspection-enum"
 import type { SortType } from "@/_types/sort-type"
-import { WorksSeriesAddButton } from "@/routes/($lang).dashboard._index/_components/works-series-add-button"
+import { WorksSeriesAddButton } from "@/routes/($lang).my._index/_components/works-series-add-button"
 import { useSuspenseQuery } from "@apollo/client/index"
 import { graphql } from "gql.tada"
 import { useContext } from "react"
