+++ app/routes/($lang).my._index/_components/albums-list-container.tsx
@@ -4,8 +4,8 @@ import { useContext } from "react"
 import { ResponsivePagination } from "@/_components/responsive-pagination"
 import { toDateTimeText } from "@/_utils/to-date-time-text"
 import { useSuspenseQuery } from "@apollo/client/index"
-import { AlbumsList } from "@/routes/($lang).dashboard._index/_components/albums-list"
-import { WorksSeriesAddButton } from "@/routes/($lang).dashboard._index/_components/works-series-add-button"
+import { AlbumsList } from "@/routes/($lang).my._index/_components/albums-list"
+import { WorksSeriesAddButton } from "@/routes/($lang).my._index/_components/works-series-add-button"
 import type { IntrospectionEnum } from "@/_lib/introspection-enum"
 import { partialAlbumFieldsFragment } from "@/_graphql/fragments/partial-album-fields"
 import { partialUserFieldsFragment } from "@/_graphql/fragments/partial-user-fields"
