+++ app/routes/($lang).my._index/_components/bookmark-list-container.tsx
@@ -1,0 +1,174 @@
+import { AuthContext } from "@/_contexts/auth-context"
+import { useContext } from "react"
+import { toDateTimeText } from "@/_utils/to-date-time-text"
+import { BookmarkWorksList } from "@/routes/($lang).my._index/_components/bookmark-works-list"
+import { ResponsivePagination } from "@/_components/responsive-pagination"
+import { useQuery } from "@apollo/client/index"
+import { partialWorkFieldsFragment } from "@/_graphql/fragments/partial-work-fields"
+import { graphql } from "gql.tada"
+
+type Props = {
+  page: number
+  maxCount: number
+  setPage: (page: number) => void
+}
+
+/**
+ * ブックマーク一覧コンテナ
+ */
+export const BookmarkListContainer = (props: Props) => {
+  const authContext = useContext(AuthContext)
+
+  if (
+    authContext.isLoading ||
+    authContext.isNotLoggedIn ||
+    authContext.userId === undefined ||
+    !authContext.userId
+  ) {
+    return null
+  }
+  const { data: userResp, refetch } = useQuery(userQuery, {
+    skip: authContext.isLoading,
+    variables: {
+      bookmarksOffset: 16 * props.page,
+      bookmarksLimit: 16,
+      userId: decodeURIComponent(authContext.userId),
+      worksWhere: {},
+      followeesWorksWhere: {},
+      followersWorksWhere: {},
+      worksOffset: 0,
+      worksLimit: 0,
+      followeesOffset: 0,
+      followeesLimit: 0,
+      followeesWorksOffset: 0,
+      followeesWorksLimit: 0,
+      followersOffset: 0,
+      followersLimit: 0,
+      followersWorksOffset: 0,
+      followersWorksLimit: 0,
+    },
+  })
+
+  const works = userResp?.user?.bookmarkWorks ?? []
+
+  return (
+    <>
+      <BookmarkWorksList
+        works={works.map((work) => ({
+          id: work.id,
+          title: work.title,
+          thumbnailImageUrl: work.smallThumbnailImageURL,
+          likesCount: work.likesCount,
+          bookmarksCount: work.bookmarksCount ?? 0,
+          commentsCount: work.commentsCount ?? 0,
+          viewsCount: work.viewsCount,
+          createdAt: toDateTimeText(work.createdAt), // Convert createdAt to string
+          accessType: work.accessType,
+          isTagEditable: work.isTagEditable,
+        }))}
+      />
+      <div className="mt-4 mb-8">
+        <ResponsivePagination
+          perPage={16}
+          maxCount={props.maxCount}
+          currentPage={props.page}
+          onPageChange={(page: number) => {
+            props.setPage(page)
+          }}
+        />
+      </div>
+    </>
+  )
+}
+
+const userQuery = graphql(
+  `query User(
+    $userId: ID!,
+    $worksOffset: Int!,
+    $worksLimit: Int!,
+    $worksWhere: UserWorksWhereInput,
+    $followeesOffset: Int!,
+    $followeesLimit: Int!,
+    $followeesWorksOffset: Int!,
+    $followeesWorksLimit: Int!,
+    $followeesWorksWhere: UserWorksWhereInput,
+    $followersOffset: Int!,
+    $followersLimit: Int!,
+    $followersWorksOffset: Int!,
+    $followersWorksLimit: Int!
+    $followersWorksWhere: UserWorksWhereInput,
+    $bookmarksOffset: Int!,
+    $bookmarksLimit: Int!,
+    $bookmarksWhere: UserWorksWhereInput,
+  ) {
+    user(id: $userId) {
+      id
+      biography
+      createdBookmarksCount
+      login
+      nanoid
+      name
+      receivedLikesCount
+      receivedViewsCount
+      awardsCount
+      followCount
+      followersCount
+      worksCount
+      iconUrl
+      headerImageUrl
+      webFcmToken
+      isFollower
+      isFollowee
+      headerImageUrl
+      works(offset: $worksOffset, limit: $worksLimit, where: $worksWhere) {
+        ...PartialWorkFields
+      }
+      followees(offset: $followeesOffset, limit: $followeesLimit) {
+        id
+        name
+        iconUrl
+        headerImageUrl
+        biography
+        isFollower
+        isFollowee
+        enBiography
+        works(offset: $followeesWorksOffset, limit: $followeesWorksLimit, where: $followeesWorksWhere) {
+          ...PartialWorkFields
+        }
+      }
+      followers(offset: $followersOffset, limit: $followersLimit) {
+        id
+        name
+        iconUrl
+        headerImageUrl
+        biography
+        isFollower
+        isFollowee
+        enBiography
+        works(offset: $followersWorksOffset, limit: $followersWorksLimit, where: $followersWorksWhere) {
+          ...PartialWorkFields
+        }
+      }
+      bookmarkWorks(offset: $bookmarksOffset, limit: $bookmarksLimit, where: $bookmarksWhere) {
+        ...PartialWorkFields
+      }
+      featuredSensitiveWorks {
+        ...PartialWorkFields
+      }
+      featuredWorks {
+        ...PartialWorkFields
+      }
+      biography
+      enBiography
+      instagramAccountId
+      twitterAccountId
+      githubAccountId
+      siteURL
+      mailAddress
+      promptonUser {
+        id
+      }
+    }
+  }`,
+  [partialWorkFieldsFragment],
+)
