+++ app/routes/($lang).my._index/_components/create-album-dialog.tsx
@@ -6,7 +6,7 @@ import { CropImageField } from "@/_components/crop-image-field"
 import { Input } from "@/_components/ui/input"
 import { AutoResizeTextarea } from "@/_components/auto-resize-textarea"
 import { Button } from "@/_components/ui/button"
-import { SelectCreatedWorksDialog } from "@/routes/($lang).dashboard._index/_components/select-created-works-dialog"
+import { SelectCreatedWorksDialog } from "@/routes/($lang).my._index/_components/select-created-works-dialog"
 import { AppLoadingPage } from "@/_components/app/app-loading-page"
 import { ScrollArea } from "@/_components/ui/scroll-area"
 import { AuthContext } from "@/_contexts/auth-context"
