+++ app/routes/($lang).my._index/_components/my-home-content-container.tsx
@@ -9,10 +9,8 @@ type Props = {
 export const DashboardHomeContentContainer = (props: Props) => {
   return (
     <>
-      <div className="h-auto rounded-md bg-zinc-50 md:h-[640px] dark:bg-zinc-700">
-        <div className="w-full rounded-t-md bg-zinc-100 p-4 font-bold dark:bg-zinc-800">
-          {props.title}
-        </div>
+      <div className="h-auto rounded-md md:h-[640px] ">
+        <div className="w-full rounded-t-md p-4 font-bold">{props.title}</div>
         {props.children}
       </div>
     </>
