+++ app/routes/($lang).my._index/_types/content-type.ts
