+++ app/routes/($lang).settings.profile/_components/setting-profile-form.tsx
@@ -4,7 +4,7 @@ import { CropImageField } from "@/_components/crop-image-field"
 import { Button } from "@/_components/ui/button"
 import { Separator } from "@/_components/ui/separator"
 import { AuthContext } from "@/_contexts/auth-context"
-import { SelectCreatedWorksDialog } from "@/routes/($lang).dashboard._index/_components/select-created-works-dialog"
+import { SelectCreatedWorksDialog } from "@/routes/($lang).my._index/_components/select-created-works-dialog"
 import { useQuery, useSuspenseQuery } from "@apollo/client/index"
 import { graphql } from "gql.tada"
 import { Loader2Icon, Pencil, PlusIcon } from "lucide-react"
